ai zombie,fixed movement

+++ HeroeZombie/src/heroezombie/EnemigoBasico.java
@@ -42,12 +42,12 @@ public class EnemigoBasico extends GameObject{
     
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int)x,(int)y, 16, 16);
     }
 
     
     public Rectangle getBounds() {
-        return new Rectangle(x,y,16,16);
+        return new Rectangle((int)x,(int)y,16,16);
     }
     
 }
