+++ HeroeZombie/src/heroezombie/HUD.java
@@ -14,9 +14,9 @@ import java.awt.Graphics;
  */
 public class HUD {
     
-    public static int vida = 100;
+    public static float vida = 100;
     
-    private int greenValue = 0;
+    private float greenValue = 0;
     
     private int score = 0;
     
@@ -24,7 +24,7 @@ public class HUD {
     
     public void tick(){
         vida= Juego.clamp(vida, 0, 100);
-        greenValue = Juego.clamp(greenValue, 0, 255);
+        greenValue =Juego.clamp(greenValue, 0, 255);
         greenValue= vida*2;
         score++;
     }
@@ -36,8 +36,8 @@ public class HUD {
     public void render(Graphics g){
         g.setColor(Color.red);
         g.fillRect(15, 15,200 , 16);
-        g.setColor(new Color (75, greenValue, 0));
-        g.fillRect(15, 15,vida*2 , 16);
+        g.setColor(new Color (75,(int)greenValue, 0));
+        g.fillRect(15, 15, (int) (vida*2), 16);
         g.setColor(Color.white);
         g.drawRect(15, 15,200 , 16);
         g.drawString("Puntuación: " + score, 15 , 64);
